refactor: change imports after change types directory name
